Refactor CMakeLists.txt and include Rapidfuzz library in vcpkg.json
Updated `vcpkg.json` file to include the `Rapidfuzz` dependency and modified the `CMakeLists.txt` to handle the changes. The modification included the removal of the `FetchContent` for Rapidfuzz and relying on the packages declared in `vcpkg.json` instead. Also adjusted the way OpenSSL libraries are handled for a streamlined and cleaner approach.